requirements.txt
Django==2.1.2
djangorestframework==3.9.0
mysqlclient==1.3.14
Pillow==5.4.1
